delone triangulation

+++ go_lib/src/intersect.cpp
@@ -98,7 +98,7 @@ namespace go{
         while (verts.size() > 3) {
             bool ear_found = false;
 
-            for (int i = 0; i < verts.size(); ++i) {
+            for (size_t i = 0; i < verts.size(); ++i) {
                 if (is_ear(verts, i)) {
                     int prev = (i - 1 + verts.size()) % verts.size();
                     int next = (i + 1) % verts.size();
